Add sfx and music, add Texture type, animations
Texture type added which has texture id, width, height, channel count.
It is useful for passing to render functions which require the dims
of a texture, like a texture atlas for sprite animations.

The groundwork for sprite animations is laid down. Will have to
consider how it will tie into the entity system, if at all.