Remove extraneous files, add sprite animations
Sprite Animations are stored in their own array and contain texture and
frame data to play an animation.

One of the two following strategies will be employed next:

Magic number:

Entites can have an animation playing by setting their animation_id to
a valid id.

If an entity has animation_id set to 0xdeadbeef, it will be considered
to have a static sprite and skip the sprite animation updates.

Every sprite is animated:

Even sprites with no animations will use a Sprite Animation, set to a
single frame on loop.